Transform special characters in DevTools package
